spaces between brackets in quizpage declaration and rename of quiz folder
